requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
chromadb>=0.4.22
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
ragas>=0.1.0
datasets>=2.14.6
nltk>=3.8.1
pandas>=2.0.3
tqdm>=4.66.1
PyYAML>=6.0.1
